feat: correct component modal kick player
