+++ src/components/KickPlayerModal/KickPlayerModal.scss
